implemented a complete mandelbrot visualization on a separate branch

+++ src/processing/sketches/Cnum.java
@@ -2,32 +2,32 @@ package processing.sketches;
 
 // a + bi
 public class Cnum {
-    private float a;
-    private float b;
+    private double a;
+    private double b;
 
-    public Cnum(float a, float b) {
+    public Cnum(double a, double b) {
         this.a = a;
         this.b = b;
     }
 
     public Cnum() {
-        this.a = 0.0f;
-        this.b = 0.0f;
+        this.a = 0.0;
+        this.b = 0.0;
     }
 
-    public float getA() {
+    public double getA() {
         return a;
     }
 
-    public void setA(float a) {
+    public void setA(double a) {
         this.a = a;
     }
 
-    public float getB() {
+    public double getB() {
         return b;
     }
 
-    public void setB(float b) {
+    public void setB(double b) {
         this.b = b;
     }
 }
